refactor: flatten app layout
